Removed QueriesPollingContext
