+++ src/api/react-query-wrappers/useMutationWrapper.ts
@@ -1,9 +1,11 @@
 import React from 'react'
 import { useDialog, useLoadingOverlay, useToastNotification } from '@/contexts'
-import { useMutation } from '@tanstack/react-query'
+import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { UseMutationWrapper } from '../comon.api.types'
 import { useJwt } from '@/hooks/useJwt'
-import { useQueriesPolling } from '@/contexts/QueriesPollingContext'
+import { ExponentialBackoffTimeout } from './react-query-wrappers.utils'
+
+let exponentialBackoffTimeout: ExponentialBackoffTimeout | undefined
 
 /**
  * This abstract and takes care of most of what's needed for the application mutations:
@@ -46,7 +48,16 @@ export const useMutationWrapper: UseMutationWrapper = (mutationFn, options) => {
   const { showOverlay, hideOverlay } = useLoadingOverlay()
   const { openDialog, closeDialog } = useDialog()
   const { hasSessionExpired } = useJwt()
-  const { requestPolling } = useQueriesPolling()
+  const queryClient = useQueryClient()
+
+  const requestPolling = React.useCallback(() => {
+    const refetchActiveQueries = () => {
+      queryClient.refetchQueries({ type: 'active' })
+    }
+
+    exponentialBackoffTimeout?.cancel()
+    exponentialBackoffTimeout = new ExponentialBackoffTimeout(refetchActiveQueries, 8)
+  }, [queryClient])
 
   /**
    * Wraps the react-query's onError option property. Handles the success toast notification and the starts refetch/polling.
