+++ src/contexts/ProvidersWrapper.tsx
@@ -6,7 +6,6 @@ import { LoadingOverlayContextProvider } from './LoadingOverlayContext'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import { queryClientConfig } from '@/config/query-client'
-import { QueriesPollingContextProvider } from './QueriesPollingContext'
 
 const queryClient = new QueryClient(queryClientConfig)
 
@@ -16,10 +15,8 @@ const ProvidersWrapper: React.FC<{ children: React.ReactNode }> = ({ children })
       <LoadingOverlayContextProvider>
         <ToastNotificationContextProvider>
           <QueryClientProvider client={queryClient}>
-            <QueriesPollingContextProvider>
-              {children}
-              <ReactQueryDevtools initialIsOpen={false} />
-            </QueriesPollingContextProvider>
+            {children}
+            <ReactQueryDevtools initialIsOpen={false} />
           </QueryClientProvider>
         </ToastNotificationContextProvider>
       </LoadingOverlayContextProvider>
